WIP: Spring 7 - Security config package moves + custom AntPathRequestMatcher
- Spring Authorization Server merged into spring-security-config (Security 7): move
  OAuth2AuthorizationServerConfiguration/Configurer imports to the new
  org.springframework.security.config.annotation.web.* packages.
- ObjectPostProcessor moved org.springframework.security.config.annotation ->
  org.springframework.security.config.
- AntPathRequestMatcher removed in Security 7 (PathPattern can't express mid-pattern
  /**/). Added a faithful drop-in org.hisp.dhis.webapi.security.AntPathRequestMatcher
  backed by Spring AntPathMatcher (servletPath+pathInfo, case-sensitive) to preserve
  exact request-authorization behaviour; swapped imports in DhisWebApiWebSecurityConfig
  + AuthenticationApiTestBase (35 call sites unchanged). NEEDS security review.
