+++ dhis-2/dhis-test-web-api/src/test/java/org/hisp/dhis/test/webapi/AuthenticationApiTestBase.java
@@ -39,7 +39,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
 import org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers;
 import org.springframework.security.web.FilterChainProxy;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
+import org.hisp.dhis.webapi.security.AntPathRequestMatcher;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.transaction.annotation.Transactional;
